+++ silo-core/contracts/incentives/lib/DistributionTypes.sol
@@ -1,16 +1,17 @@
 // SPDX-License-Identifier: agpl-3.0
+
 pragma solidity 0.8.28;
 
 library DistributionTypes {
     struct IncentivesProgramCreationInput {
         string name;
         address rewardToken;
-        uint104 emissionPerSecond;
+        uint256 emissionPerSecond;
         uint40 distributionEnd;
     }
 
     struct AssetConfigInput {
-        uint104 emissionPerSecond;
+        uint256 emissionPerSecond;
         uint256 totalStaked;
         address underlyingAsset;
     }
